refactor code, create separate LinkifiedText component
